Add script for updating Chromium version
