+++ extensions/ql-vscode/gulpfile.ts/chromium-version.json
@@ -1,0 +1,4 @@
+{
+  "chromiumVersion": "114",
+  "electronVersion": "25.8.0"
+}
